Delay import of pysss as much as possible
When importing at the top level, it makes the build fail (build server does
not have the module) with pybuilder 0.11.9 even though the module is not even
tested. Pybuilder 0.11.9 seems to be the last version that still works with
our Python 2.6 setup, so we cannot update to 0.11.10 which does not have the
problem.